pyproject.toml
[build-system]
requires = ["setuptools>=67", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "contro"
version = "0.1.0"
description = "Open-source headless CMS framework built with Django."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
  "Django>=5.0,<6.0",
  "djangorestframework>=3.14",
  "djangorestframework-simplejwt>=5.3",
  "graphene-django>=3.1",
  "django-filter>=23.3",
  "django-cors-headers>=4.3",
  "django-environ>=0.11",
  "psycopg2-binary>=2.9",
  "Pillow>=10.0",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["contro*"]
